requirements.txt
spotipy==2.21.0
requests>=2.31.0
tidalapi==0.7.2
pyyaml==6.0
tqdm==4.64.1